requirements.txt
peewee
pymysql
flask
flask-script
flask-wtf
flask-login
mysql-connector==2.1.4
SQLAlchemy
python-crontab
psutil
pysmb
pyasn1
twisted